Integrate TestJvmProcess
